html files can be opened... ouga bouga happy

+++ src/html_file.h
@@ -1,8 +1,10 @@
 #ifndef HTML_FILE_H
 #define HTML_FILE_H
 
+#include <stddef.h>
+
 struct html_file {
-  long filesize;
+  size_t filesize;
   char *file_buffer;
 };
